Update Japanese teatre.po (#193)
* Update theatre.po

* added Japanese translation

For the translation of some words, I priority used the words used to be used in Flipnote Hatena.
(e.g. Follow→お気に入りに登録(add to favorite), original Fripnote→親作品(parent's work) etc)

I couldn't come up with a good translation of "Hapy Flipnoting!", so I didn't translate it.
(“Happy うごライフ！” (うごライフ : Flipnote+life) is one of the translations I thought of, but I also think it's not bad to not to be translated. I'd like to hear the opinion from other translator...)

* added Japanese translation

For the translation of some words, I priority used the words used to be used in Flipnote Hatena.
(e.g. Follow→お気に入りに登録(add to favorite), original Fripnote→親作品(parent's work) etc)

I couldn't come up with a good translation of "Hapy Flipnoting!", so I didn't translate it.
(“Happy うごライフ！” (うごライフ : Flipnote+life) is one of the translations I thought of, but I also think it's not bad to not to be translated. I'd like to hear the opinion from other translator...)

* Update teatre.po

Fixed some translations.