Fixed Dutch Translation (#214)
* Fix grammatical errors

Was this file machine translated? A lot of the errors seem like they've come from an automatic English -> Dutch translation, mostly stumbling on the lack of gender neutral pronouns in Dutch, and confusing it for plural.

* Remove @ symbols

As recommended by the "How to contribute" file

* Fix grammatical errors

Fun fact: Incorrectly added spaces are usually called "The English disease" in The Netherlands, because of the fact that Dutch sometimes glues words together where English would leave them separate.

* Update channelList.po

* Rename "commentaar" to "reacties"

Comments, as in comments on a YouTube video, are called reacties here. So from an English point of view it would make sense to call replies "reacties" and comments "commentaar", it makes more sense in Dutch to call both of them reacties.

* Fix grammatical errors

* Fix grammatical errors

+ rename commentaar to reactie

* Fix grammatical errors

* Fix grammatical errors

* Fix grammatical errors

* Fix grammatical errors

+ rename commentaar to reacties

* Fix grammatical error

* Fix grammatical mistakes

+ rename commentaar to reacties

* Fix grammatical errors

* Fix grammatical error

* Fix grammatical errors

* Add POTENTIAL_CONTRIBUTER_NOTE translation

+ fix grammatical error

* Fix grammatical errors

* Translate mail.po to Dutch

* Fix grammatical errors

* rename soort to type

* Translate mail_content_removed.po

* Translate mail_content_restored.po to Dutch

* Translate mail_report_accepted.po to Dutch

* Translate mail_spinoff.po to Dutch

* Fix grammatical errors + remove @ signs

* Fix grammatical errors

* Fix grammatical errors

* Fix grammatical errors

* Fix grammatical errors

* Translate mail_report_rejected.po to Dutch

* Fix grammatical errors

* Fix grammatical errors

* Fix grammatical errors

* Fix grammatical errors

* Fix grammatical errors + translate remaining strings

* Fix grammatical errors

* Fix grammatical errors + commentaar -> reacties

* Commentaar -> reacties

* Fix grammatical error

* Translate trophies.po to Dutch

* Change double quotes to single quotes

Both because of syntax problems, and to be consistent with other strings